# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Exoplanet Travel Agency API
  version: 0.1.0
  description: |
    REST API for browsing exoplanet destinations and booking travel holidays.

    ## Conventions
    - All responses are JSON (`application/json`) unless otherwise stated.
    - Pagination uses `page` (1-indexed) and `pageSize`.
    - Error responses follow a consistent shape:
      `{ "error": { "code": string, "message": string, "details"?: any } }`
  contact:
    name: API Maintainer
    email: your-email@example.com
  license:
    name: Academic Use
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://YOUR-DEPLOYED-URL-HERE
    description: Deployed server (optional)
...
paths:
...
  /api/bookings/{id}:
    get:
      tags: [Bookings]
      operationId: getBooking
      summary: Get a single booking by ID
      security:
        - cookieAuth: []
      description: |
        Returns a single booking with user and planet details.

        - Only the **booking owner** or an **ADMIN** can view a booking.
        - Returns `403 Forbidden` if the user is neither owner nor admin.
        - Requires a valid session cookie.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags: [Bookings]
      operationId: updateBooking
      summary: Update a booking (partial)
      security:
        - cookieAuth: []
      description: |
        Partially updates a booking.

        - Only the **booking owner** or an **ADMIN** can update a booking.
        - Returns `403 Forbidden` if the user is neither owner nor admin.
        - Requires a valid session cookie.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingRequest"
            examples:
              example:
                summary: Update travel class
                value:
                  travelClass: First Class (Warp Drive)
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Bookings]
      operationId: deleteBooking
      summary: Delete a booking
      security:
        - cookieAuth: []
      description: |
        Deletes a booking.

        - Only the **booking owner** or an **ADMIN** can delete a booking.
        - Returns `403 Forbidden` if the user is neither owner nor admin.
        - Requires a valid session cookie.
        - Returns `204 No Content` on success.
      parameters:
        - $ref: "#/components/parameters/IdPathParam"
      responses:
        "204":
          description: Deleted successfully (no content)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: exo-session
      description: |
        JWT session token stored in an httpOnly cookie named `exo-session`.
        Obtain by calling POST /api/auth/login or POST /api/auth/register.
  parameters:
    IdPathParam:
      in: path
      name: id
      required: true
      description: Resource id (Prisma `cuid()`).
      schema:
        type: string
        minLength: 10
      example: cmlp9k8qn0003dskufekwmztt
...
  responses:
    BadRequest:
      description: Validation error (invalid parameters or body)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example:
              value:
                error:
                  code: VALIDATION_ERROR
                  message: Invalid query parameters
                  details:
                    fieldErrors:
                      page:
                        - Too small: expected number to be >=1
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example:
              value:
                error:
                  code: NOT_FOUND
                  message: Resource not found
    Unauthorized:
      description: Authentication required (no valid session cookie)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example:
              value:
                error:
                  code: UNAUTHORIZED
                  message: Authentication required
    Forbidden:
      description: Insufficient permissions (not owner or admin)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            example:
              value:
                error:
                  code: FORBIDDEN
                  message: You can only modify your own bookings
  schemas:
...
    Booking:
      type: object
      required: [id, userId, planetId, travelClass, status, bookingDate]
      properties:
        id:
          type: string
          description: Unique identifier (Prisma cuid).
        userId:
          type: string
          description: User id who owns the booking.
        planetId:
          type: string
          description: Exoplanet id being booked.
        travelClass:
          type: string
          description: Travel class label chosen by the user.
        status:
          type: string
          description: Booking status (e.g. CONFIRMED, CANCELLED).
        bookingDate:
          type: string
          format: date-time
          description: Timestamp when the booking was created.
...
    UpdateBookingRequest:
      type: object
      description: Provide at least one field to update.
      properties:
        travelClass:
          type: string
          description: Updated travel class selection.
        status:
          type: string
          description: Updated booking status.
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error message.
            details:
              nullable: true
              description: Optional validation or debug details.
...